youtube-assistant/requirements.txt
langchain
openai
streamlit
python-dotenv
numexpr
youtube-transcript-api
faiss-cpu